Back/frs_back/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==6.0.0
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
comm==0.1.3
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.0
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-bootstrap-v5==1.0.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
executing==1.2.0
Faker==8.1.1
fastjsonschema==2.16.3
Flask==2.3.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fqdn==1.5.1
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
importlib-metadata==2.1.1
inflection==0.5.1
ipykernel==6.22.0
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.0
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
locust==2.15.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.3
mistune==2.0.5
msgpack==1.0.5
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
npm==0.1.1
numpy==1.24.3
oauthlib==3.2.2
optional-django==0.1.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.10.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.6
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
ruamel.yaml==0.17.21
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.2.1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.1
widgetsnbextension==4.0.7
zipp==3.4.1
zope.event==4.6
zope.interface==6.0
